Move passwordless resource
