+++ workos/passwordless.py
@@ -3,7 +3,7 @@ import workos
 from workos.types.passwordless.passwordless_session_type import PasswordlessSessionType
 from workos.utils.http_client import SyncHTTPClient
 from workos.utils.request_helper import REQUEST_METHOD_POST
-from workos.resources.passwordless import PasswordlessSession
+from workos.types.passwordless.passwordless_session import PasswordlessSession
 from workos.utils.validation import Module, validate_settings
